refactor: 重构组件名称
