+++ src/BootstrapBlazor.Server/Components/Samples/Table/TablesFilter.razor
@@ -65,7 +65,7 @@
             </TableColumn>
             <TableColumn @bind-Field="@context.Count" Width="100" Sortable="true" Filterable="true">
                 <FilterTemplate>
-                    <TableColumnFilter TFilter="CustomerFilter"></TableColumnFilter>
+                    <Filter TFilter="CustomerFilter"></Filter>
                 </FilterTemplate>
             </TableColumn>
         </TableColumns>
@@ -211,22 +211,22 @@
             <TableColumn @bind-Field="@context.DateTime" Width="180" Sortable="true" />
             <TableColumn @bind-Field="@context.Name" Width="100" Sortable="true" Filterable="true">
                 <FilterTemplate>
-                    <TableColumnFilter TFilter="MultiFilter" FilterParameters="_multiFilterParameter1"></TableColumnFilter>
+                    <Filter TFilter="MultiFilter" FilterParameters="_multiFilterParameter1"></Filter>
                 </FilterTemplate>
             </TableColumn>
             <TableColumn @bind-Field="@context.Address" Sortable="true" Filterable="true">
                 <FilterTemplate>
-                    <TableColumnFilter TFilter="MultiFilter" FilterParameters="_multiFilterParameter2"></TableColumnFilter>
+                    <Filter TFilter="MultiFilter" FilterParameters="_multiFilterParameter2"></Filter>
                 </FilterTemplate>
             </TableColumn>
             <TableColumn @bind-Field="@context.Complete" Width="100" Sortable="true" Filterable="true">
                 <FilterTemplate>
-                    <TableColumnFilter TFilter="MultiFilter" FilterParameters="_multiFilterParameter3"></TableColumnFilter>
+                    <Filter TFilter="MultiFilter" FilterParameters="_multiFilterParameter3"></Filter>
                 </FilterTemplate>
             </TableColumn>
             <TableColumn @bind-Field="@context.Education" Width="100" Sortable="true" Filterable="true">
                 <FilterTemplate>
-                    <TableColumnFilter TFilter="MultiFilter" FilterParameters="_multiFilterParameter4"></TableColumnFilter>
+                    <Filter TFilter="MultiFilter" FilterParameters="_multiFilterParameter4"></Filter>
                 </FilterTemplate>
             </TableColumn>
             <TableColumn @bind-Field="@context.Count" Width="150" Sortable="true" />
