+++ src/BootstrapBlazor/Components/Filters/Filter.razor
@@ -1,0 +1,28 @@
+﻿@namespace BootstrapBlazor.Components
+@inherits ComponentBase
+@typeparam TFilter
+
+@if (_isHeaderRow)
+{
+    @RenderFilter()
+}
+else
+{
+    <div class="card filter-item">
+        <div class="card-header"><span>@Title</span></div>
+        <div class="card-body">
+            @RenderFilter()
+        </div>
+        <div class="card-footer">
+            <div class="d-flex flex-fill">
+                @if (ShowMoreButton)
+                {
+                    <Button Color="Color.None" OnClick="OnClickPlus" Icon="@PlusIcon" IsDisabled="@(_count == 1)"></Button>
+                    <Button Color="Color.None" OnClick="OnClickMinus" Icon="@MinusIcon" IsDisabled="@(_count == 0)"></Button>
+                }
+            </div>
+            <Button Color="Color.None" class="filter-dismiss" OnClick="OnClickReset" Text="@ClearButtonText"></Button>
+            <Button Color="Color.None" class="filter-dismiss" OnClick="OnClickConfirm" Text="@FilterButtonText"></Button>
+        </div>
+    </div>
+}
