+++ src/BootstrapBlazor/wwwroot/scss/components.scss
@@ -43,7 +43,7 @@
 @import "../../Components/FlipClock/FlipClock.razor.scss";
 @import "../../Components/FileIcon/FileIcon.razor.scss";
 @import "../../Components/Filters/MultiFilter.razor.scss";
-@import "../../Components/Filters/TableFilter.razor.scss";
+@import "../../Components/Filters/TableColumnFilter.razor.scss";
 @import "../../Components/Footer/Footer.razor.scss";
 @import "../../Components/FullScreen/FullScreenButton.razor.scss";
 @import "../../Components/GoTop/GoTop.razor.scss";
